Continue to use v2 shortcut if CONDA_PYTHON_EXE does not exist.
V1 shortcuts do not work with without this environment variable (e.g. Micromamba), since a base environment is not known. The shortcut is created, but is broken. It is better to not create a shortcut.

If there is an error, or v2 shortcut is broken, then we will deal with that issue when it arises.